New_/IAS-Project-Group-5-dev-jasika/model_manager/requirements.txt
azure-storage-file-share
Flask==2.0.2
Flask_PyMongo==2.3.0
numpy==1.22.2
pymongo==4.1.0
Werkzeug==2.0.3
